+++ src/ideal/internal/translations/en_US.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  idealSelectYourBank: 'Select Your Bank',
+  idealBack: 'Back',
+  idealContinuingTo: 'Continuing to'
+};