+++ src/ideal/internal/translations/nl_NL.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  idealSelectYourBank: 'Selecteer uw bank',
+  idealBack: 'Terug',
+  idealContinuingTo: 'Doorgaan naar'
+};