+++ src/ideal/static/sandbox-approval.js
@@ -1,11 +1,9 @@
 'use strict';
 
 var querystring = require('../../lib/querystring');
-var sanitizeHtml = require('../../lib/sanitize-html');
 var isWhitelistedDomain = require('../../lib/is-whitelisted-domain');
 
-var TEXT_PARAMS = ['amount', 'currency', 'final_status'];
-var REQUIRED_PARAMS = TEXT_PARAMS.concat('redirect_url');
+var REQUIRED_PARAMS = ['redirect_url'];
 
 function verifyParams(data) {
   REQUIRED_PARAMS.forEach(function (param) {
@@ -21,12 +19,6 @@ function verifyParams(data) {
 
 function populateDom(data) {
   var link;
-
-  TEXT_PARAMS.forEach(function (param) {
-    var domNode = document.getElementById(param);
-
-    domNode.textContent = sanitizeHtml(data[param]);
-  });
 
   link = document.getElementById('redirect_url');
   link.href = data.redirect_url;