+++ tasks/frame-service.js
@@ -1,8 +1,11 @@
 
 'use strict';
 
+var del = require('del');
+var fs = require('fs');
 var gulp = require('gulp');
 var rename = require('gulp-rename');
+var run = require('run-sequence');
 var replace = require('gulp-replace');
 var browserify = require('./browserify');
 var minifyHTML = require('gulp-minifier');
@@ -10,11 +13,13 @@ var VERSION = require('../package.json').version;
 
 var DIST_DIR = 'dist/hosted/web/' + VERSION + '/';
 var JS_TASKS = [];
+var JS_DELETE_TASKS = [];
 var HTML_TASKS = [];
 var FRAMES = ['dispatch'];
 
 FRAMES.forEach(function (frame) {
   var jsTaskName = 'build:frame-service:js:' + frame + '-frame';
+  var jsDeleteTaskName = 'build:frame-service:js:delete' + frame + '-frame';
   var htmlTaskName = 'build:frame-service:html:' + frame + '-frame';
 
   gulp.task(jsTaskName, function (done) {
@@ -22,14 +27,18 @@ FRAMES.forEach(function (frame) {
       standalone: 'frameService',
       main: 'src/lib/frame-service/internal/dispatch-frame/index.js',
       out: 'frame-service-' + frame + '-frame.js',
-      dist: DIST_DIR + 'js'
+      dist: DIST_DIR + 'js',
+      uglify: false
     }, done);
   });
 
   JS_TASKS.push(jsTaskName);
 
   gulp.task(htmlTaskName, function () {
+    var jsFile = fs.readFileSync(DIST_DIR + 'js/frame-service-dispatch-frame.js');
+
     return gulp.src('src/lib/frame-service/internal/dispatch-frame/index.html')
+      .pipe(replace('@BUILT_FILE', jsFile))
       .pipe(rename(function (path) {
         path.basename = 'dispatch-frame';
       }))
@@ -49,8 +58,19 @@ FRAMES.forEach(function (frame) {
   });
 
   HTML_TASKS.push(htmlTaskName);
+
+  gulp.task(jsDeleteTaskName, function () {
+    var jsFilePath = DIST_DIR + 'js/frame-service-dispatch-frame.js';
+
+    return del(jsFilePath);
+  });
+
+  JS_DELETE_TASKS.push(jsDeleteTaskName);
 });
 
 gulp.task('build:frame-service:html', HTML_TASKS);
 gulp.task('build:frame-service:js', JS_TASKS);
-gulp.task('build:frame-service', ['build:frame-service:js', 'build:frame-service:html']);
+gulp.task('build:frame-service:js:delete', JS_DELETE_TASKS);
+gulp.task('build:frame-service', function (done) {
+  run('build:frame-service:js', 'build:frame-service:html', 'build:frame-service:js:delete', done);
+});