+++ tasks/ideal.js
@@ -1,15 +1,19 @@
 'use strict';
 
+var del = require('del');
 var gulp = require('gulp');
+var fs = require('fs');
 var rename = require('gulp-rename');
 var replace = require('gulp-replace');
+var run = require('run-sequence');
 var browserify = require('./browserify');
 var minify = require('gulp-minifier');
 var VERSION = require('../package.json').version;
 
 var DIST_DIR = 'dist/hosted/web/' + VERSION + '/';
 var STATIC_DIR = 'dist/hosted/web/static/';
-var JS_TASKS = ['build:ideal:js:external'];
+var JS_TASKS = [];
+var DELETE_INTERNAL_JS_TASKS = [];
 var HTML_TASKS = [];
 var CSS_TASKS = [];
 var IMG_TASKS = [];
@@ -18,12 +22,13 @@ var internalFrames = ['issuers', 'redirect'];
 var staticFrames = ['sandbox-approval'];
 
 internalFrames.forEach(function (frame) {
-  var taskNames = ['js', 'css', 'html'].map(function (lang) {
-    return 'build:ideal:' + lang + ':' + frame + '-frame';
+  var taskNames = ['js', 'css', 'html', 'js:delete'].map(function (subtask) {
+    return 'build:ideal:frame:' + subtask + ':' + frame + '-frame';
   });
   var jsTaskName = taskNames[0];
   var cssTaskName = taskNames[1];
   var htmlTaskName = taskNames[2];
+  var deleteInternalJSTaskName = taskNames[3];
 
   JS_TASKS.push(jsTaskName);
   gulp.task(jsTaskName, function (done) {
@@ -31,7 +36,8 @@ internalFrames.forEach(function (frame) {
       standalone: 'braintree.ideal',
       main: 'src/ideal/internal/' + frame + '-frame.js',
       out: 'ideal-' + frame + '-frame.js',
-      dist: DIST_DIR + 'js'
+      dist: DIST_DIR + 'js',
+      uglify: false
     }, done);
   });
 
@@ -56,7 +62,10 @@ internalFrames.forEach(function (frame) {
 
   HTML_TASKS.push(htmlTaskName);
   gulp.task(htmlTaskName, function () {
+    var jsFile = fs.readFileSync(DIST_DIR + 'js/ideal-' + frame + '-frame.js');
+
     return gulp.src('src/ideal/internal/' + frame + '-frame.html')
+      .pipe(replace('@BUILT_FILE', jsFile))
       .pipe(rename(function (path) {
         path.basename = 'ideal-' + path.basename;
       }))
@@ -73,15 +82,23 @@ internalFrames.forEach(function (frame) {
       }))
       .pipe(gulp.dest(DIST_DIR + 'html'));
   });
+
+  DELETE_INTERNAL_JS_TASKS.push(deleteInternalJSTaskName);
+  gulp.task(deleteInternalJSTaskName, function () {
+    var jsFilePath = DIST_DIR + 'js/ideal-' + frame + '-frame.js';
+
+    return del(jsFilePath);
+  });
 });
 
 staticFrames.forEach(function (frame) {
-  var taskNames = ['js', 'img', 'html'].map(function (lang) {
+  var taskNames = ['js', 'img', 'html', 'js:delete'].map(function (lang) {
     return 'build:static-ideal:' + lang + ':' + frame;
   });
   var jsTaskName = taskNames[0];
   var imgTaskName = taskNames[1];
   var htmlTaskName = taskNames[2];
+  var deleteInternalJSTaskName = taskNames[3];
 
   JS_TASKS.push(jsTaskName);
   gulp.task(jsTaskName, function (done) {
@@ -89,7 +106,8 @@ staticFrames.forEach(function (frame) {
       standalone: 'braintree.ideal',
       main: 'src/ideal/static/' + frame + '.js',
       out: 'ideal-' + frame + '.js',
-      dist: STATIC_DIR + 'js'
+      dist: STATIC_DIR + 'js',
+      uglify: false
     }, done);
   });
 
@@ -101,7 +119,10 @@ staticFrames.forEach(function (frame) {
 
   HTML_TASKS.push(htmlTaskName);
   gulp.task(htmlTaskName, function () {
+    var jsFile = fs.readFileSync(STATIC_DIR + 'js/ideal-'+ frame + '.js');
+
     return gulp.src('src/ideal/static/' + frame + '.html')
+      .pipe(replace('@BUILT_FILE', jsFile))
       .pipe(rename(function (path) {
         path.basename = 'ideal-' + path.basename;
       }))
@@ -114,9 +135,19 @@ staticFrames.forEach(function (frame) {
       }))
       .pipe(gulp.dest(STATIC_DIR + 'html'));
   });
+
+  DELETE_INTERNAL_JS_TASKS.push(deleteInternalJSTaskName);
+  gulp.task(deleteInternalJSTaskName, function () {
+    // All JavaScript built files for static frames
+    // are not needed after inlining.
+    // Delete the static js folder.
+    var jsFolder = STATIC_DIR + 'js/';
+
+    return del(jsFolder);
+  });
 });
 
-gulp.task('build:ideal:js:external', function (done) {
+gulp.task('build:ideal:js', function (done) {
   browserify({
     standalone: 'braintree.ideal',
     main: 'src/ideal/index.js',
@@ -125,8 +156,12 @@ gulp.task('build:ideal:js:external', function (done) {
   }, done);
 });
 
-gulp.task('build:ideal:html', HTML_TASKS);
-gulp.task('build:ideal:js', JS_TASKS);
-gulp.task('build:ideal:css', CSS_TASKS);
-gulp.task('build:ideal:img', IMG_TASKS);
-gulp.task('build:ideal', ['build:ideal:js', 'build:ideal:css', 'build:ideal:html', 'build:ideal:img']);
+gulp.task('build:ideal:frame:html', HTML_TASKS);
+gulp.task('build:ideal:frame:js', JS_TASKS);
+gulp.task('build:ideal:frame:css', CSS_TASKS);
+gulp.task('build:ideal:frame:img', IMG_TASKS);
+gulp.task('build:ideal:frame:js:delete', DELETE_INTERNAL_JS_TASKS);
+gulp.task('build:ideal:frame', ['build:ideal:frame:css', 'build:ideal:frame:img'], function (done) {
+  run('build:ideal:frame:js', 'build:ideal:frame:html', 'build:ideal:frame:js:delete', done);
+});
+gulp.task('build:ideal', ['build:ideal:js', 'build:ideal:frame']);