+++ tasks/masterpass.js
@@ -1,34 +1,44 @@
 'use strict';
 
+var del = require('del');
+var fs = require('fs');
 var gulp = require('gulp');
 var rename = require('gulp-rename');
 var replace = require('gulp-replace');
+var run = require('run-sequence');
 var browserify = require('./browserify');
 var minify = require('gulp-minifier');
 var VERSION = require('../package.json').version;
 
 var DIST_DIR = 'dist/hosted/web/' + VERSION + '/';
-var JS_TASKS = ['build:masterpass:js:external'];
+var JS_TASKS = [];
 var HTML_TASKS = [];
+var DELETE_INTERNAL_JS_TASKS = [];
 var FRAMES = ['redirect', 'loading', 'landing'];
 
 FRAMES.forEach(function (frame) {
-  var jsTaskName = 'build:masterpass:js:' + frame + '-frame';
-  var htmlTaskName = 'build:masterpass:html:' + frame + '-frame';
+  var jsTaskName = 'build:masterpass:frame:js:' + frame + '-frame';
+  var htmlTaskName = 'build:masterpass:frame:html:' + frame + '-frame';
+  var deleteInternalJSTaskName = 'build:masterpass:frame:js:delete:' + frame + '-frame';
 
   gulp.task(jsTaskName, function (done) {
     browserify({
       standalone: 'braintree.masterpass',
       main: 'src/masterpass/internal/' + frame + '-frame.js',
       out: 'masterpass-' + frame + '-frame.js',
-      dist: DIST_DIR + 'js'
+      dist: DIST_DIR + 'js',
+      uglify: false
     }, done);
   });
 
   JS_TASKS.push(jsTaskName);
 
   gulp.task(htmlTaskName, function () {
+    var jsFilePath = DIST_DIR + 'js/masterpass-' + frame + '-frame';
+    var jsFile = fs.readFileSync(jsFilePath + '.js');
+
     return gulp.src('src/masterpass/internal/frame.html')
+      .pipe(replace('@BUILT_FILE', jsFile))
       .pipe(rename(function (path) {
         path.basename = 'masterpass-' + frame + '-' + path.basename;
       }))
@@ -48,9 +58,17 @@ FRAMES.forEach(function (frame) {
   });
 
   HTML_TASKS.push(htmlTaskName);
+
+  gulp.task(deleteInternalJSTaskName, function () {
+    var jsFilePath = DIST_DIR + 'js/masterpass-' + frame + '-frame';
+
+    return del(jsFilePath + '.js');
+  });
+
+  DELETE_INTERNAL_JS_TASKS.push(deleteInternalJSTaskName);
 });
 
-gulp.task('build:masterpass:js:external', function (done) {
+gulp.task('build:masterpass:js', function (done) {
   browserify({
     standalone: 'braintree.masterpass',
     main: 'src/masterpass/index.js',
@@ -59,6 +77,10 @@ gulp.task('build:masterpass:js:external', function (done) {
   }, done);
 });
 
-gulp.task('build:masterpass:html', HTML_TASKS);
-gulp.task('build:masterpass:js', JS_TASKS);
-gulp.task('build:masterpass', ['build:masterpass:js', 'build:masterpass:html']);
+gulp.task('build:masterpass:frame:html', HTML_TASKS);
+gulp.task('build:masterpass:frame:js', JS_TASKS);
+gulp.task('build:masterpass:frame:js:delete', DELETE_INTERNAL_JS_TASKS);
+gulp.task('build:masterpass:frame', function (done) {
+  run('build:masterpass:frame:js', 'build:masterpass:frame:html', 'build:masterpass:frame:js:delete', done);
+});
+gulp.task('build:masterpass', ['build:masterpass:js', 'build:masterpass:frame']);