+++ tasks/paypal.js
@@ -1,34 +1,44 @@
 'use strict';
 
+var del = require('del');
+var fs = require('fs');
 var gulp = require('gulp');
 var rename = require('gulp-rename');
 var replace = require('gulp-replace');
+var run = require('run-sequence');
 var browserify = require('./browserify');
 var minifyHTML = require('gulp-minifier');
 var VERSION = require('../package.json').version;
 
 var DIST_DIR = 'dist/hosted/web/' + VERSION + '/';
-var JS_TASKS = ['build:paypal:js:external'];
+var JS_TASKS = [];
+var JS_DELETE_TASKS = [];
 var HTML_TASKS = [];
 var FRAMES = ['redirect', 'cancel'];
 
 FRAMES.forEach(function (frame) {
-  var jsTaskName = 'build:paypal:js:' + frame + '-frame';
-  var htmlTaskName = 'build:paypal:html:' + frame + '-frame';
+  var jsTaskName = 'build:paypal:frame:js:' + frame + '-frame';
+  var jsDeleteTaskName = 'build:paypal:frame:js:delete:' + frame + '-frame';
+  var htmlTaskName = 'build:paypal:frame:html:' + frame + '-frame';
+
 
   gulp.task(jsTaskName, function (done) {
     browserify({
       standalone: 'braintree.paypal',
       main: 'src/paypal/internal/' + frame + '-frame.js',
       out: 'paypal-' + frame + '-frame.js',
-      dist: DIST_DIR + 'js'
+      dist: DIST_DIR + 'js',
+      uglify: false
     }, done);
   });
 
   JS_TASKS.push(jsTaskName);
 
   gulp.task(htmlTaskName, function () {
+    var jsFile = fs.readFileSync(DIST_DIR + 'js/paypal-' + frame + '-frame.js', 'utf8');
+
     return gulp.src('src/paypal/internal/frame.html')
+      .pipe(replace('@BUILT_FILE', jsFile))
       .pipe(rename(function (path) {
         path.basename = 'paypal-' + frame + '-' + path.basename;
       }))
@@ -48,10 +58,19 @@ FRAMES.forEach(function (frame) {
   });
 
   HTML_TASKS.push(htmlTaskName);
+
+  gulp.task(jsDeleteTaskName, function () {
+    var jsFilePath = DIST_DIR + 'js/paypal-' + frame + '-frame.js'
+
+    return del(jsFilePath);
+  });
+
+  JS_DELETE_TASKS.push(jsDeleteTaskName);
+
 });
 
-HTML_TASKS.push('build:paypal:html:landing-frame');
-gulp.task('build:paypal:html:landing-frame', function () {
+HTML_TASKS.push('build:paypal:frame:html:landing-frame');
+gulp.task('build:paypal:frame:html:landing-frame', function () {
   return gulp.src('src/paypal/internal/landing-frame.html')
     .pipe(rename('paypal-landing-frame.html'))
     .pipe(gulp.dest(DIST_DIR + 'html'))
@@ -66,7 +85,7 @@ gulp.task('build:paypal:html:landing-frame', function () {
     .pipe(gulp.dest(DIST_DIR + 'html'));
 });
 
-gulp.task('build:paypal:js:external', function (done) {
+gulp.task('build:paypal:js', function (done) {
   browserify({
     standalone: 'braintree.paypal',
     main: 'src/paypal/index.js',
@@ -75,6 +94,10 @@ gulp.task('build:paypal:js:external', function (done) {
   }, done);
 });
 
-gulp.task('build:paypal:html', HTML_TASKS);
-gulp.task('build:paypal:js', JS_TASKS);
-gulp.task('build:paypal', ['build:paypal:js', 'build:paypal:html']);
+gulp.task('build:paypal:frame:html', HTML_TASKS);
+gulp.task('build:paypal:frame:js', JS_TASKS);
+gulp.task('build:paypal:frame:js:delete', JS_DELETE_TASKS);
+gulp.task('build:paypal:frame', function (done) {
+  run('build:paypal:frame:js', 'build:paypal:frame:html', 'build:paypal:frame:js:delete', done);
+});
+gulp.task('build:paypal', ['build:paypal:js', 'build:paypal:frame']);