+++ tasks/three-d-secure.js
@@ -1,8 +1,11 @@
 'use strict';
 
+var del = require('del');
+var fs = require('fs');
 var gulp = require('gulp');
 var rename = require('gulp-rename');
 var replace = require('gulp-replace');
+var run = require('run-sequence');
 var browserify = require('./browserify');
 var minifyHTML = require('gulp-minifier');
 var VERSION = require('../package.json').version;
@@ -11,10 +14,13 @@ var DIST_DIR = 'dist/hosted/web/' + VERSION + '/';
 var FRAMES = ['bank', 'authentication-complete'];
 
 var HTML_TASKS = FRAMES.map(function (frame) {
-  var htmlTaskName = 'build:three-d-secure:html:' + frame + '-frame';
+  var htmlTaskName = 'build:three-d-secure:frame:html:' + frame + '-frame';
 
   gulp.task(htmlTaskName, function () {
+    var jsFile = fs.readFileSync(DIST_DIR + 'js/three-d-secure-frame.js');
+
     return gulp.src('src/three-d-secure/internal/' + frame + '-frame.html')
+      .pipe(replace('@BUILT_FILE', jsFile))
       .pipe(rename(function (path) {
         path.basename = 'three-d-secure-' + path.basename;
       }))
@@ -36,16 +42,17 @@ var HTML_TASKS = FRAMES.map(function (frame) {
   return htmlTaskName;
 });
 
-gulp.task('build:three-d-secure:js:internal', function (done) {
+gulp.task('build:three-d-secure:frame:js', function (done) {
   browserify({
     standalone: 'braintree.three-d-secure',
     main: 'src/three-d-secure/internal/index.js',
     out: 'three-d-secure-frame.js',
-    dist: DIST_DIR + 'js'
+    dist: DIST_DIR + 'js',
+    uglify: false
   }, done);
 });
 
-gulp.task('build:three-d-secure:js:external', function (done) {
+gulp.task('build:three-d-secure:js', function (done) {
   browserify({
     standalone: 'braintree.three-d-secure',
     main: 'src/three-d-secure/index.js',
@@ -54,6 +61,15 @@ gulp.task('build:three-d-secure:js:external', function (done) {
   }, done);
 });
 
-gulp.task('build:three-d-secure:html', HTML_TASKS);
-gulp.task('build:three-d-secure:js', ['build:three-d-secure:js:internal', 'build:three-d-secure:js:external']);
-gulp.task('build:three-d-secure', ['build:three-d-secure:js', 'build:three-d-secure:html']);
+gulp.task('build:three-d-secure:frame:js:delete', function (done) {
+  var frameJsFilePath = DIST_DIR + 'js/three-d-secure-frame.js';
+
+  return del(frameJsFilePath);
+});
+
+gulp.task('build:three-d-secure:frame:html', HTML_TASKS);
+gulp.task('build:three-d-secure:frame', function (done) {
+  run('build:three-d-secure:frame:js', 'build:three-d-secure:frame:html', 'build:three-d-secure:frame:js:delete', done);
+});
+
+gulp.task('build:three-d-secure', ['build:three-d-secure:frame', 'build:three-d-secure:js']);